refactored matrix to reside in own source and header
